+++ pages/ctf/tsukuctf2021/[id].tsx
@@ -1,31 +1,20 @@
 import { Layout } from "../../../components/Layout";
 import { getAllPostIds, getPostData } from "../../../lib/posts";
 import { GetStaticProps, GetStaticPaths } from "next";
-
-import ReactMarkdown from "react-markdown";
-import path from "path";
+import MarkdownRender, {
+  PostPropType,
+} from "../../../components/MarkdownRender";
 
 import { useRouter } from "next/router";
 import { useState, useEffect } from "react";
 
-// "tsukuctf2021";
+const ctfId = "tsukuctf2021";
 
-type PostPropType = {
-  postData: {
-    title: string;
-    description: string;
-    author: string;
-    genre: string;
-    solver: number;
-    point: number;
-    contentHtml: string;
-  };
-};
 const Post = ({ postData }: PostPropType) => {
   const { title, description, author, genre, solver, point, contentHtml } =
     postData;
   const router = useRouter();
-  const [ctfId, setCTFId] = useState<string>();
+  const [ctfId, setCTFId] = useState<string>("");
   useEffect(() => {
     if (router.asPath !== router.route) {
       setCTFId(String(router.pathname).split("/").slice(-2, -1)[0]);
@@ -33,20 +22,7 @@ const Post = ({ postData }: PostPropType) => {
   }, [router]);
   return (
     <Layout title={`${postData.title}-${ctfId}`}>
-      <ReactMarkdown
-        components={{
-          img: ({ node, ...props }) => {
-            const src = path.join(
-              `https://raw.githubusercontent.com/SecHack365-Fans/SecHack365-Fans.github.io/master/src/pages/ctf/${ctfId}/writeups/`,
-              title,
-              props.src as string
-            );
-            return <img src={src} alt={props.alt} style={{ width: "80%" }} />;
-          },
-        }}
-      >
-        {postData.contentHtml}
-      </ReactMarkdown>
+      <MarkdownRender id={ctfId} title={title} content={contentHtml} />
     </Layout>
   );
 };
@@ -54,7 +30,7 @@ const Post = ({ postData }: PostPropType) => {
 export default Post;
 
 export const getStaticPaths: GetStaticPaths = async () => {
-  const paths = getAllPostIds("tsukuctf2021");
+  const paths = getAllPostIds(ctfId);
   return {
     paths,
     fallback: false,
@@ -62,7 +38,7 @@ export const getStaticPaths: GetStaticPaths = async () => {
 };
 
 export const getStaticProps: GetStaticProps = async ({ params }) => {
-  const postData = await getPostData(params?.id as string, "tsukuctf2021");
+  const postData = await getPostData(params?.id as string, ctfId);
   return {
     props: {
       postData,
